breaks server into multiple parts, begins adding sqlite db
